[anhthv] fix: Update DTO assessmentOption

+++ backend-api/backend-api/Models/DTOs/AssessmentOptionDTO.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public string? OptionText { get; set; }
-        public int Point { get; set; }
+        public double Point { get; set; }
     }
 }
